backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]
bcrypt==3.2.0
fastapi-limiter
redis
pyotp
qrcode
Pillow